KeyboardService: reduce CircularBuffer size
The online IDE for mbed classic forces us to use s130 SoftDevice, which
reduces the available RAM. With 512 bytes of circular buffer, the
example doesn't work properly, so we reduce it to 256 bytes.

This is unfortunate, because it drastically limits the amount of data we
can prepare. But other build systems, that allow to use s110, also make
it easy to override the KEYBUFFER_SIZE macro.