Remove HIDDeviceInformationService
This class was not the right way to implement PnPID: it should be added
as an argument for DeviceInformationService. For the moment, we aim to
simplify BLE_HID as much as possible for merging into BLE API.

Warning: this commit breaks MacOS 10.10 support! We used a hack hidden
inside HIDDeviceInformationService because, for some reason, MacOS
needed a characteristic external to HIDService to have security
requirements, or it just didn't want to pair. I don't have access to
Apple hardware anymore, so I have no idea if this got fixed.

To use the library with MacOS 10.10, apply "requireSecurity(NO_MITM)"
onto an external characteristic if your first attempt doesn't work. 10.9
and below are not expected to work at all currently, but maybe you will
get lucky.